DashBoard/locker-dashboard-1/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
pandas==1.5.3
SQLAlchemy==1.4.39
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.20.0